+++ app/models/festivity_market_page.rb
@@ -1,3 +1,5 @@
 class FestivityMarketPage < FestivityBasePage
 
+  include Concerns::FestivityArtistMethods
+
 end
